suffix imported version names with ".mvn".
if later one wants to publish artifacts for a project directly, it should be possible to use the version name.  a project already imported from maven as "v1.1.mvn" might choose to later use "v1.1" for the library binary by itself and "v1.1-docs" to separately publish the documentation, for example, while still leaving "v1.1.mvn" published including both blobs in the same folder plus the pom file for those interested in it.