make a new implementation of a version discoverer that completely ignores the maven metadata files that (theoretically) are a programmatically valid list of versions, because it turns out *maven completely ignores the maven metadata files*, maven central performs zero checks that they match reality, and in fact many projects published on maven central contain metadata files that are just flat wrong.
i've known for a while now that there's no programmatic API to list files in a release on maven, and that wound is half-healed if still sore... but now i'm finding that there's no way to programmatically list versions either?  unfuckingbelievable.  this is literally inferior to an unauthenticated ftp server.